Create hackrooms model, views, controller and routes
This creates a model for hackrooms so we can save information about them and
present them on a page for people to understand them.

The description should be used to describe the language you're focused on, and
the long description can tell members what they should expect from the
hackroom, what they need to do to prepare for them (setup their environment,
bring a laptop etc).

It will also display a list of people who are interested in this language.

We can add other features to this later to build on it to make it easier to
facilitate and advertise these sessions.