Install CanCanCan for abilities
CanCanCan (known as CanCan as it was originally called before someone else
picked it up) is a permissions system that makes it really easy to secure the
app and only allow people to do what they should be able to do.

We'll specify what normal users are allowed to do, then create an admin role
where we can list the things admins can do.