Create API route namespace and base controller
This will allow us to create API controllers in `app/controllers/api/` that
will inherit from a controller we can use to put any code that applies to all
API controllers, such as authentication to prevent unauthorized actions.

Using `ActionController::API` instead of ApplicationController gets rid of all
the features we won't need in an API controller, but we can add any stripped
features that we do need.