Add authentication methods to API controller
`request_authenticated?` will check if the request URL contains the WLC API
key and return true if it does, false if not.

You can use this to run an if statement to check if a request is authorized or
not.

`respond_404_unauthorized` will return an unauthorized response if you'd like
to deny access to an endpoint that hasn't been sent with the API key.